Move simyan.comicvine to simyan.service
