+++ docs/simyan/service.md
@@ -1,0 +1,4 @@
+# Service
+
+::: simyan.service.Comicvine
+::: simyan.service.ComicvineResource
