+++ tests/conftest.py
@@ -7,19 +7,17 @@ import os
 
 import pytest
 
-from simyan.comicvine import Comicvine
+from simyan.service import Comicvine
 from simyan.sqlite_cache import SQLiteCache
 
 
 @pytest.fixture(scope="session")
-def comicvine_api_key():
+def api_key() -> str:
     """Set the ComicVine API key fixture."""
     return os.getenv("COMICVINE__API_KEY", default="IGNORED")
 
 
 @pytest.fixture(scope="session")
-def session(comicvine_api_key) -> Comicvine:
+def session(api_key: str) -> Comicvine:
     """Set the Simyan session fixture."""
-    return Comicvine(
-        api_key=comicvine_api_key, cache=SQLiteCache("tests/cache.sqlite", expiry=None)
-    )
+    return Comicvine(api_key=api_key, cache=SQLiteCache("tests/cache.sqlite", expiry=None))
