+++ tests/test_characters.py
@@ -7,9 +7,9 @@ from datetime import datetime
 
 import pytest
 
-from simyan.comicvine import Comicvine, ComicvineResource
 from simyan.exceptions import ServiceError
 from simyan.schemas.character import CharacterEntry
+from simyan.service import Comicvine, ComicvineResource
 
 
 def test_character(session: Comicvine):
